rag/requirements.txt
torch
transformers
chromadb
TTS @ git+https://github.com/coqui-ai/TTS.git
pyyaml
openai
pdfplumber
python-docx
